fix(pipeline): drop .int() on ranked id to satisfy Anthropic outputFormat
Zod's z.number().int() lowers to JSON Schema integer with implicit
safe-int min/max bounds, which Anthropic's structured output endpoint
rejects: "For 'integer' type, properties maximum, minimum are not
supported". Use plain z.number() — IDs are still numeric and validated
downstream.

+++ packages/pipeline/src/processors/rank.ts
@@ -47,7 +47,7 @@ export interface RankResult {
 }
 
 const rankedEntrySchema = z.object({
-  id: z.number().int(),
+  id: z.number(),
   score: z.number(),
   rationale: z.string().min(1),
   summary: z.string(),
